requirements.txt
pandas==2.2.3
aiohttp==3.11.11
apache-airflow==2.10.4
ensure==1.0.2
python-box==6.0.2
numpy==1.25.0
lightgbm==4.5.0
seaborn==0.13.2
statsmodels==0.14.4
xgboost==2.1.3
matplotlib==3.10.0
scikit-learn==1.6.0
scipy==1.15.0
prophet==1.1.6
dash==2.18.2
fastapi==0.115.6
uvicorn==0.34.0
tensorflow==2.16.1
python-dotenv==1.0.1

-e .
